loligo-back/requirements.txt
anyio==3.7.0
arrow==1.3.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
certifi==2023.5.7
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cookiecutter==2.5.0
cycler==0.11.0
fastapi==0.97.0
fonttools==4.38.0
h11==0.14.0
httptools==0.6.1
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.3
mdurl==0.1.2
MechanicalSoup==1.3.0
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.2.0
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pydantic==1.10.9
Pygments==2.17.2
pyparsing==3.0.9
pysolar==0.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2022.7.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
schedule==1.2.0
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
text-unidecode==1.3
types-python-dateutil==2.8.19.14
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
asgiref==3.4.1
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
fastapi==0.68.0
h11==0.12.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.15.0
